refactor(AddQuote): rename AddQuote in QuoteForm
